pyproject.toml
[project]
name = "httpie-csrf"
version = "1.0.0"
description = "CSRF plugin for HTTPie."
authors = [
    { name = "Xdynix", email = "Lizard.rar@gmail.com" },
]
dependencies = [
    "httpie<4.0.0,>=3.2.0",
    "requests<3.0.0,>=2.28.0",
    "python-decouple>=3.8",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.entry-points."httpie.plugins.transport.v1"]
httpie_csrf = "httpie_csrf.plugin:HttpieCsrfPlugin"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
lint = [
    "pre-commit>=3.7.1",
    "mypy>=1.10.1",
    "types-requests>=2.32.0.20240622",
]

[tool.pdm.scripts.post_install]
cmd = ["pre-commit", "install"]
help = "install pre-commit hooks"

[tool.pdm.scripts.lint]
cmd = ["pre-commit", "run", "--all-files"]
env = { PYTHONUTF8 = "1" }
help = "execute linters"

[tool.ruff]
cache-dir = ".cache/ruff"
target-version = "py310"

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "I", # isort
    "UP", # pyupgrade
    "S", # flake8-bandit
    "B", # flake8-bugbear
    "C", # flake8-comprehensions
    "DJ", # flake8-django
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "FURB", # refurb
    "RUF", # Ruff-specific rules
]
ignore = [
    # Suppressed by ruff formatter.
    "W191", # tab-indentation
    "E111", # indentation-with-invalid-multiple
    "E114", # indentation-with-invalid-multiple-comment
    "E117", # over-indented
]

[tool.mypy]
strict = true
cache_dir = ".cache/mypy"

[[tool.mypy.overrides]]
module = "httpie.plugins"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "decouple"
ignore_missing_imports = true
